Fixed LiteralExpression.Render's handling of negative numbers. Reworked NumberParser.cs's parsing methods to take ReadOnlySpan<char> instead of string. Fixed various validations of arguments that should be numeric that were actually checking specifically if the expression's type was STRING. Added property DataTpe.IsNumeric. Fixed the output of negative INTEGER and LONG values in PrintEmitter.cs. Added support for the graphics WINDOW: - Renamed Rect.cs, Point.cs and PointExtensions.cs to IntegerRect.cs, IntegerPoint.cs and IntegerPointExtensions.cs. - Added new Point.cs with float values. Added PointExtensions to provide tuple deconstruction as with IntegerPoint. - Added class Window.cs to describe the graphics window and provide translation helpers. - Added a Window field to GraphicsLibrary.cs. Updated RefreshParameters to reset it to the identity. - Changed the type of LastPoint from IntegerPoint to Point. - Changed all but the primary overload of Graphics.PixelSet to use float coordinates, and added a new overload matching the primary overload but with float coordinates that translates them through Window and then updates LastPoint. - Changed all but the primary overload of Graphics.Line/LineTo to use float coordinates, and added a new overload matching the primary overload but with float coordinates that translates them through Window and then updates LastPoint. - Changed all but the primary overload of Graphics.Box/BoxTo to use float coordinates, and added a new overload matching the primary overload but with float coordinates that translates them through Window and then updates LastPoint. - Changed all but the primary overload of Graphics.FillBox/FillBoxTo to use float coordinates, and added a new overload matching the primary overload but with float coordinates that translates them through Window and then updates LastPoint. - Changed all but the primary overload of Graphics.Ellipse to use float coordinates, and added a new overload matching the primary overload but with float coordinates that translates them through Window and then updates LastPoint. Updated PixelSetStatement.cs to use SINGLE coordinates. Added support for CIRCLE statements: - Added ExecutionEngine type CircleStatement.cs. - Updated TranslateStatement in Compiler.cs to convert CodeModel CircleStatements to ExecutionEngine CircleStatements. Added an overload of CompilerException.TypeMismatch that takes the context Token directly.

+++ QBX/Firmware/PointExtensions.cs
@@ -2,7 +2,7 @@
 
 public static class PointExtensions
 {
-	public static void Deconstruct(this Point pt, out int x, out int y)
+	public static void Deconstruct(this Point pt, out float x, out float y)
 	{
 		x = pt.X;
 		y = pt.Y;
